Hadoop comparison results adjusted to common interface

+++ src/main/java/mapreduceinjava8/wordcount/WordCountHadoopResultsImporter.java
@@ -10,23 +10,22 @@ public class WordCountHadoopResultsImporter {
 
     private static final String TABULATOR = "\t";
 
-    public Map<String, Integer> importResults(final String file) {
+    public Map<String, Long> importResults(final String file) {
 
         final List<String> lines = FileAccessor.readLinesFromResources(file);
 
-        final Map<String, Integer> results = Maps.newHashMap();
+        final Map<String, Long> results = Maps.newHashMap();
 
         for (final String line : lines) {
 
             final String[] fields = line.split(TABULATOR);
 
             final String word = fields[0];
-            final Integer count = Integer.parseInt(fields[1]);
+            final Long count = Long.parseLong(fields[1]);
 
             results.put(word, count);
         }
 
         return results;
     }
-
 }
